Add dynamic step elision
This optimization allows Knit to avoid rebuilding a file if, dynamically
during the build, it determines that the prereq has not changed. For
example, if you change a source file by modifying a comment, Knit will
re-run the compiler but may notice that the object file is unchanged and
therefore will not invoke the linker or the rest of the build. This
feature is only enabled when using hash-based detection.

When using the "steps" printer, you may notice that the build may jump
some steps. This is because those steps have been dynamically elided by
this optimization, but the total step counter that runs before the build
cannot know which steps might be elided.